+++ src/RiskManagement/__init__.py
@@ -17,6 +17,7 @@ from .HoldingTimeoutRule import HoldingTimeoutRule
 from .PositionAnomalyRule import PositionAnomalyRule
 from .PairDrawdownRule import PairDrawdownRule
 from .RiskManager import RiskManager
+from .RiskHandler import RiskHandler
 
 __all__ = [
     'RiskRule',
@@ -26,5 +27,6 @@ __all__ = [
     'HoldingTimeoutRule',
     'PositionAnomalyRule',
     'PairDrawdownRule',
-    'RiskManager'
+    'RiskManager',
+    'RiskHandler'
 ]